:art: Refactors StatPing widget
